Exception Handling

+++ src/main/java/com/jose30a2/dscommerce/services/ProductService.java
@@ -2,11 +2,16 @@ package com.jose30a2.dscommerce.services;
 
 import com.jose30a2.dscommerce.dto.ProductDTO;
 import com.jose30a2.dscommerce.entities.Product;
+import com.jose30a2.dscommerce.exceptions.DataBaseException;
+import com.jose30a2.dscommerce.exceptions.ResourceNotFoundException;
 import com.jose30a2.dscommerce.repositories.ProductRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -21,10 +26,20 @@ public class ProductService {
     //  @Transactional de Spring
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
+        /** Exige usar try catch
         Optional<Product> result = repository.findById(id);
         Product product = result.get();
         ProductDTO dto = new ProductDTO(product);
-        return dto;
+            // Alternativo
+
+         Product product = repository.findById(id).get();
+         ProductDTO dto = new ProductDTO(product);
+         */
+
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso no encontrado: " + id));
+
+        return new ProductDTO(product);
     }
 
     @Transactional(readOnly = true)
@@ -46,20 +61,31 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
+        try{
+            // Objeto que no e buscado no repositorio solo es referenciado por el id
+            Product entity = repository.getReferenceById(id);
 
-        // Objeto que no e buscado no repositorio solo es referenciado por el id
-        Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
 
-        copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
 
-        entity = repository.save(entity);
+            return new ProductDTO(entity);
 
-        return new ProductDTO(entity);
+        }catch(EntityNotFoundException e){
+            throw new ResourceNotFoundException("recurso nao encontrado");
+        }
+
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if( !repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso nao encontrado");
+        }try{
+            repository.deleteById(id);
+        }catch(DataIntegrityViolationException e ){
+            throw new DataBaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
